sushi orders now ordered correctly
